EES 4809: Upgrade to NextJS v14 (#4785)
NextJS in @frontend upgrade to v14; NextJS removed from @common; refactoring of some components, babel replaced with swc in @common and @frontend
