+++ src/explore-education-statistics-admin/__mocks__/reactLeafletMock.ts
@@ -1,0 +1,3 @@
+module.exports = {
+  MapContainer: () => {},
+};
