+++ src/explore-education-statistics-admin/src/components/ButtonLink.tsx
@@ -2,11 +2,6 @@ import classNames from 'classnames';
 import React, { ReactNode } from 'react';
 import { Link as RouterLink, LinkProps } from 'react-router-dom';
 
-import {
-  SetCommonButtonLink,
-  ButtonLinkType,
-} from '@common/components/ButtonLink';
-
 type Props = {
   children: ReactNode;
   className?: string;
@@ -31,7 +26,5 @@ const ButtonLink = ({ children, className, to, variant, ...props }: Props) => {
     </RouterLink>
   );
 };
-
-SetCommonButtonLink(ButtonLink as ButtonLinkType);
 
 export default ButtonLink;
