--- src/explore-education-statistics-common/.babelrc
@@ -1,21 +1,0 @@
-{
-  "presets": [
-    [
-      "next/babel",
-      {
-        "preset-env": {
-          "useBuiltIns": "usage",
-          "corejs": 3,
-          "targets": {
-            "browsers": ["ie>=11", "safari>=12", "ios>=9.3", "last 1 version"]
-          }
-        }
-      }
-    ]
-  ],
-  "env": {
-    "test": {
-      "plugins": ["./babel-url-import-meta-plugin.js"]
-    }
-  }
-}
