+++ src/explore-education-statistics-common/package.json
@@ -11,7 +11,6 @@
     "@radix-ui/react-dialog": "1.0.4",
     "@tanstack/react-query": "4.32.0",
     "axios": "^1.4.0",
-    "babel-preset-react-app": "^9.1.2",
     "classnames": "^2.3.2",
     "color2k": "^2.0.2",
     "core-js": "^3.31.1",
@@ -30,14 +29,12 @@
     "lodash": "^4.17.21",
     "memoizee": "^0.4.14",
     "nanoid": "^4.0.0",
-    "next": "^12.3.4",
     "qs": "^6.11.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-helmet": "^6.1.0",
     "react-highlight-words": "^0.20.0",
     "react-hook-form": "7.43.9",
-    "react-lazyload":"github:twobin/react-lazyload#f1faffda816180f39d3ff8cf3d9c2060f8d9d623",
+    "react-lazyload": "github:twobin/react-lazyload#f1faffda816180f39d3ff8cf3d9c2060f8d9d623",
     "react-leaflet": "^4.2.1",
     "react-markdown": "^8.0.0",
     "recharts": "^2.7.2",
@@ -49,8 +46,8 @@
     "yup": "^1.3.3"
   },
   "devDependencies": {
-    "@swc/core": "^1.3.71",
-    "@swc/jest": "^0.2.27",
+    "@swc/core": "^1.4.17",
+    "@swc/jest": "^0.2.36",
     "@testing-library/dom": "^9.3.4",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^14.2.1",
@@ -65,13 +62,11 @@
     "@types/qs": "^6.9.7",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
-    "@types/react-helmet": "^6.1.0",
     "@types/react-highlight-words": "^0.16.4",
     "@types/react-lazyload": "^3.2.3",
     "@types/react-modal": "^3.16.0",
     "@types/sanitize-html": "^2.9.0",
     "@types/uuid": "^8.3.0",
-    "babel-jest": "^29.6.1",
     "eslint": "^8.43.0",
     "jest": "^29.6.1",
     "jest-junit": "^16.0.0",
