+++ src/explore-education-statistics-frontend/__mocks__/reactLeafletMock.ts
@@ -1,0 +1,3 @@
+module.exports = {
+  MapContainer: () => {},
+};
