+++ src/explore-education-statistics-frontend/package.json
@@ -23,8 +23,7 @@
     "leaflet": "^1.9.4",
     "lodash": "^4.17.21",
     "memoizee": "^0.4.14",
-    "next": "^12.3.4",
-    "next-transpile-modules": "^9.1.0",
+    "next": "^14.2.2",
     "nookies": "^2.5.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
@@ -38,10 +37,8 @@
     "yup": "^1.3.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/polyfill": "^7.11.5",
-    "@swc/core": "^1.3.71",
-    "@swc/jest": "^0.2.27",
+    "@swc/core": "^1.4.17",
+    "@swc/jest": "^0.2.36",
     "@testing-library/dom": "^9.3.4",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^14.2.1",
@@ -50,16 +47,12 @@
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.3",
     "@types/leaflet": "^1.5.17",
-    "@types/lodash": "^4.14.195",
     "@types/memoizee": "^0.4.4",
     "@types/node": "^16.18.31",
     "@types/qs": "^6.9.7",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
     "@types/whatwg-fetch": "0.0.33",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-jest": "^29.6.1",
-    "babel-plugin-named-asset-import": "^0.3.6",
     "case-sensitive-paths-webpack-plugin": "^2.4.0",
     "eslint": "^8.42.0",
     "eslint-webpack-plugin": "^4.0.1",
