+++ src/explore-education-statistics-frontend/src/modules/find-statistics/components/PublicationReleaseHeadlinesSection.tsx
@@ -1,7 +1,7 @@
 import Tabs from '@common/components/Tabs';
 import TabsSection from '@common/components/TabsSection';
 import ContentBlockRenderer from '@common/modules/find-statistics/components/ContentBlockRenderer';
-import DataBlockTabs from '@common/modules/find-statistics/components/DataBlockTabs';
+import DataBlockTabs from '@frontend/modules/find-statistics/components/DataBlockTabs';
 import KeyStat, {
   KeyStatContainer,
 } from '@common/modules/find-statistics/components/KeyStat';
